Refactor (rating)/* :Split UI and Create custom hooks and create interface and put in the features folder
