+++ features/rating/components/DepartmentRating/index.ts
@@ -1,0 +1,6 @@
+export { ActionButtons } from './ActionButtons';
+export { CommentSection } from './CommentSection';
+export { DepartmentRatingUI } from './DepartmentRatingUI';
+export { ProgressBar } from './ProgressBar';
+export { QuestionDots } from './QuestionDots';
+export { RatingCard } from './RatingCard';
